requirements.txt
discord.py[voice]
google-cloud-texttospeech
google-api-python-client
youtube_dl
